requirements.txt
aiogram>=3.5.0
FastAPI==0.103.2
uvicorn==0.23.2
SQLAlchemy==2.0.21
asyncpg==0.28.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
apscheduler==3.10.4
loguru==0.7.2
tenacity==8.2.3
python-slugify==8.0.1
requests==2.31.0
httpx==0.25.0
cryptography
yookassa
alembic
psycopg2-binary
pydantic>=2.0.0

# Testing
pytest
pytest-asyncio
pytest-mock
aiosqlite